+++ com.unity.render-pipelines.high-definition/Runtime/Lighting/LightLoop/LightLoop.cs
@@ -640,7 +640,12 @@ namespace UnityEngine.Rendering.HighDefinition
         int m_CurrentShadowSortedSunLightIndex = -1;
         HDAdditionalLightData m_CurrentSunLightAdditionalLightData;
         DirectionalLightData m_CurrentSunLightDirectionalLightData;
-        internal Light GetCurrentSunLight() { return m_CurrentSunLight; }
+
+        /// <summary>
+        /// Main directional Light for the HD Render Pipeline.
+        /// </summary>
+        /// <returns>The main directional Light.</returns>
+        public Light GetMainDirectionalLight() => m_CurrentSunLight;
 
         // Screen space shadow data
         struct ScreenSpaceShadowData
@@ -1182,7 +1187,7 @@ namespace UnityEngine.Rendering.HighDefinition
                 size = new Vector2(additionalLightData.shapeWidth, additionalLightData.shapeHeight),
                 position = light.GetPosition()
             };
-            if (lightComponent == GetCurrentSunLight())
+            if (lightComponent == GetMainDirectionalLight())
             {
                 // If this is the current sun light and volumetric cloud shadows are enabled we need to render the shadows
                 if (HasVolumetricCloudsShadows_IgnoreSun(hdCamera))