+++ com.unity.render-pipelines.high-definition/Runtime/Sky/CloudSystem/CloudLayer/CloudLayerRenderer.cs
@@ -73,7 +73,7 @@ namespace UnityEngine.Rendering.HighDefinition
             if (cloudLayer.CastShadows)
             {
                 if (m_PrecomputedData == null || m_PrecomputedData.cloudShadowsRT == null)
-                    UpdateCache(cloudLayer, HDRenderPipeline.currentPipeline.GetCurrentSunLight());
+                    UpdateCache(cloudLayer, HDRenderPipeline.currentPipeline.GetMainDirectionalLight());
 
                 cookieParams.texture = m_PrecomputedData.cloudShadowsRT;
                 cookieParams.size = new Vector2(cloudLayer.shadowSize.value, cloudLayer.shadowSize.value);